+++ core/src/pnet/mod.rs
@@ -29,7 +29,7 @@ use crypt_writer::CryptWriter;
 use log::trace;
 use rand::RngCore;
 use salsa20::{
-    stream_cipher::{NewStreamCipher, SyncStreamCipher},
+    cipher::{NewStreamCipher, SyncStreamCipher},
     Salsa20, XSalsa20,
 };
 use sha3::{digest::ExtendableOutput, Shake128};