+++ protocols/kad/src/record/store/memory.rs
@@ -80,7 +80,7 @@ impl MemoryStore {
     /// Creates a new `MemoryRecordStore` with the given configuration.
     pub fn with_config(local_id: PeerId, config: MemoryStoreConfig) -> Self {
         MemoryStore {
-            local_key: kbucket::Key::new(local_id),
+            local_key: kbucket::Key::from(local_id),
             config,
             records: HashMap::default(),
             provided: HashSet::default(),
@@ -166,9 +166,9 @@ impl<'a> RecordStore<'a> for MemoryStore {
             // It is a new provider record for that key.
             let local_key = self.local_key.clone();
             let key = kbucket::Key::new(record.key.clone());
-            let provider = kbucket::Key::new(record.provider.clone());
+            let provider = kbucket::Key::from(record.provider);
             if let Some(i) = providers.iter().position(|p| {
-                let pk = kbucket::Key::new(p.provider.clone());
+                let pk = kbucket::Key::from(p.provider);
                 provider.distance(&key) < pk.distance(&key)
             }) {
                 // Insert the new provider.
@@ -223,16 +223,16 @@ impl<'a> RecordStore<'a> for MemoryStore {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use multihash::{wrap, Code};
+    use libp2prs_core::multihash::{Code, Multihash};
     use quickcheck::*;
     use rand::Rng;
 
     fn random_multihash() -> Multihash {
-        wrap(Code::Sha2_256, &rand::thread_rng().gen::<[u8; 32]>())
+        Multihash::wrap(Code::Sha2_256.into(), &rand::thread_rng().gen::<[u8; 32]>()).unwrap()
     }
 
     fn distance(r: &ProviderRecord) -> kbucket::Distance {
-        kbucket::Key::new(r.key.clone()).distance(&kbucket::Key::new(r.provider.clone()))
+        kbucket::Key::new(r.key.clone()).distance(&kbucket::Key::from(r.provider))
     }
 
     #[test]
@@ -286,9 +286,9 @@ mod tests {
     #[test]
     fn provided() {
         let id = PeerId::random();
-        let mut store = MemoryStore::new(id.clone());
+        let mut store = MemoryStore::new(id);
         let key = random_multihash();
-        let rec = ProviderRecord::new(key, id.clone(), None);
+        let rec = ProviderRecord::new(key, id, None);
         assert!(store.add_provider(rec.clone()).is_ok());
         assert_eq!(vec![Cow::Borrowed(&rec)], store.provided().collect::<Vec<_>>());
         store.remove_provider(&rec.key, &id);
@@ -332,7 +332,7 @@ mod tests {
         let key = random_multihash();
         for _ in 0..10 {
             let prv = PeerId::random();
-            let rec = ProviderRecord::new(key.clone(), prv, None);
+            let rec = ProviderRecord::new(key, prv, None);
             let _ = store.add_provider(rec);
         }